+++ NEWS.md
@@ -1,3 +1,4 @@
+- added `check_name()` to check if elements of character vector are each a syntactically valid name
 - `check_named()` now only gives 1 warning if error = FALSE and not named
 - removed `check_tz()`